Really fix checkerboard UV coords

+++ src/materials/checkerboard.rs
@@ -20,8 +20,8 @@ impl Material for Checkerboard {
 	fn get_bxdf( &self, u : f64, v : f64 ) -> ~BxDF {
 		let u10 = ( u * 10.0 ).floor() as i64;
 		let v10 = ( v * 10.0 ).floor() as i64;
-		let uc = u * 10.0 - u10;
-		let vc = v * 10.0 - v10;
+		let uc = u * 10.0 - u10 as f64;
+		let vc = v * 10.0 - v10 as f64;
 
 		if ( u10 + v10 ) % 2 == 0 {
 			return self.mat1.get_bxdf( uc, vc );
